feat: split unit test to be closer to the code they test
